+++ fiddles/python/fiddle-0010-HttpRequest/script.py
@@ -1,0 +1,13 @@
+import os
+import urllib2
+import json
+
+
+
+def main():
+    r = urllib2.urlopen("https://bradyhouse.github.io/stockQuote/assets/json/intraday.json").read()
+    data = json.loads(r)
+    print (data)
+
+main()
+os._exit(1)